# Conflicts:
#	scripts/commands/addmission.lua
#	scripts/commands/addquest.lua
#	scripts/commands/delmission.lua
#	scripts/commands/delquest.lua
#	scripts/globals/msg.lua
#	scripts/globals/spells/bluemagic/animating_wail.lua
#	scripts/globals/spells/bluemagic/battery_charge.lua
#	scripts/globals/spells/bluemagic/mp_drainkiss.lua
#	scripts/globals/spells/curaga.lua
#	scripts/globals/spells/curaga_ii.lua
#	scripts/globals/spells/curaga_iii.lua
#	scripts/globals/spells/curaga_iv.lua
#	scripts/globals/spells/curaga_v.lua
#	scripts/globals/spells/death.lua
#	scripts/globals/spells/dispel.lua
#	scripts/zones/Sea_Serpent_Grotto/mobs/Charybdis.lua
#	src/map/enmity_container.cpp

+++ src/map/enmity_container.cpp
@@ -224,7 +224,7 @@ void CEnmityContainer::UpdateEnmityFromCure(CBattleEntity* PEntity, uint16 level
         enmity_obj->second.active = true;
     }
     else
-        m_EnmityList.emplace(PEntity->id, EnmityObject_t{ PEntity, dsp_cap(CE, 0, 30000), dsp_cap(VE, 0, 30000), true, 0 });
+        m_EnmityList.emplace(PEntity->id, EnmityObject_t{ PEntity, (int16)dsp_cap(CE, 0, 30000), (int16)dsp_cap(VE, 0, 30000), true, 0 });
 }
 
 /************************************************************************
